Set up scroll-to-hash-target rigamarole that circumvents normal hashery
By that I mean it doesn't match up the hash target #commentID with an ID
to have the browser automatically scroll down to the target (which would
be at the top of the browser, hidden by the navbar), but instead some
Javascript wizardry reads that hash and scrolls down to the correct
element with a scroll offset to accommodate the navbar.

Later on, this offset will need to be dropped to zero if the browser is
in mobile-friendly mode and there's no fixed navbar.